HYD-5785 qs-parsers: Extract to own repo
We are currently using input to QS parsing on the status screen.

There are some edge cases for parsing on the status screen that do not
apply to a general input <-> qs parser. They are things like having a
field called record_type that should read and parsed as type by the
parser.

In addition, we would like to use the parser on other areas of the GUI.
Because of this it makes sense to split out the QS parser with base
rules that apply generically to parsing input <-> qs.

  - Create initial repo.
  - Add dotfiles.
  - Extract base parsers.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>