qs-parsers: Add missing flow pragma to tokens.js
The flow pragma was missing from tokens.js. This was causing it to not
be typechecked when used as a dependency.

In turn, this led to us erroneously using inputToQsTokens as a parser
object and did catch the misuse.

By adding the flow pragma, the file with the error now typechecks
and catches the problem.

  - Add missing flow pragma and missing copyright.

Signed-off-by: Joe Grund <joe.grund@intel.com>