qs-parsers-#10 Support encoded comma and colon
Now that we are using UI-Router the query params will encode both the
comma and the colon:
, -> %2C
: -> %3A

Since this is what will be used in the string, the tokens need to be
updated to use these values instead of their ascii counterpart.
Furthermore, there may be an additional spot or two where we split on
these characters. These will need to be updated as well. A new test will
be added to test for dates and existing integration tests should now use
the new encoded format.

Fixes #10.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>